Restructure scripts directory
